Remove redundant casts
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@14380 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ mmstudio/src/org/micromanager/MainFrame.java
@@ -642,7 +642,7 @@ public class MainFrame extends JFrame implements LiveModeListener {
              comboBinning_.addItem(item);
          }
 
-         comboBinning_.setMaximumRowCount((int) items.size());
+         comboBinning_.setMaximumRowCount(items.size());
          if (items.isEmpty()) {
              comboBinning_.setEditable(true);
          } else {
